build: add @types/node
In addition to adding @types/node also update dependabot.yml to make
sure that we stay on the same major version of @types/node as the node
LTS version we are using (20 currently).

Deliberately add an older version of @types/node to see if Dependabot
does actually update to within that major version.