chore: bump package.json#packageManager
Bump package.json#packageManager to pnpm@10.12.4

The unix-dgram package has a binding.gyp file. The presence of this file
cause pnpm to suggest running `pnpm approve-builds` to approve building
of packages. The lack of an actual postinstall script in the unix-dgram
package means that nothing is actually built when the build is approved.
Because of this it was added to the ignoredBuiltDependencies list
instead.